+++ Unity.Entities.Graphics/EntitiesGraphicsCulling.cs
@@ -1,9 +1,10 @@
+// #define DISABLE_SHADOW_CULLING_CAPSULE_TEST
 // #define DISABLE_HYBRID_SPHERE_CULLING
 // #define DISABLE_HYBRID_RECEIVER_CULLING
 // #define DISABLE_INCLUDE_EXCLUDE_LIST_FILTERING
 // #define DEBUG_VALIDATE_VISIBLE_COUNTS
 // #define DEBUG_VALIDATE_COMBINED_SPLIT_RECEIVER_CULLING
-// #define DEBUG_VALIDATE_SOA_SPHERE_TEST
+// #define DEBUG_VALIDATE_VECTORIZED_CULLING
 // #define DEBUG_VALIDATE_EXTRA_SPLITS
 
 #if UNITY_EDITOR
@@ -270,10 +271,7 @@ namespace Unity.Rendering
         public UnsafeList<FrustumPlanes.PlanePacket4> ReceiverPlanePackets;
         public UnsafeList<FrustumPlanes.PlanePacket4> CombinedSplitAndReceiverPlanePackets;
         public UnsafeList<CullingSplitData> Splits;
-        public SOASphereTest SplitSOASphereTest;
-
-        public float3 LightAxisX;
-        public float3 LightAxisY;
+        public ReceiverSphereCuller ReceiverSphereCuller;
         public bool SphereTestEnabled;
 
         public static CullingSplits Create(BatchCullingContext* cullingContext, ShadowProjection shadowProjection, AllocatorManager.AllocatorHandle allocator)
@@ -318,17 +316,14 @@ namespace Unity.Rendering
             ReceiverPlanePackets = default;
             CombinedSplitAndReceiverPlanePackets = default;
             Splits = default;
-            SplitSOASphereTest = default;
-
-            LightAxisX = default;
-            LightAxisY = default;
+            ReceiverSphereCuller = default;
             SphereTestEnabled = false;
 
             // Initialize receiver planes first, so they are ready to be combined in
             // InitializeSplits
             InitializeReceiverPlanes(ref cullingContext, allocator);
             InitializeSplits(ref cullingContext, allocator);
-            InitializeSphereTest(ref cullingContext, shadowProjection, allocator);
+            InitializeSphereTest(ref cullingContext, shadowProjection);
         }
 
         private void InitializeReceiverPlanes(ref BatchCullingContext cullingContext, AllocatorManager.AllocatorHandle allocator)
@@ -480,6 +475,7 @@ namespace Unity.Rendering
                 {
                     CullingSphereCenter = p,
                     CullingSphereRadius = r,
+                    ShadowCascadeBlendCullingFactor = s.cascadeBlendCullingFactor,
                     PlanePacketOffset = planeIndex,
                     PlanePacketCount = planePackets.Length,
                     CombinedPlanePacketOffset = combinedPlaneIndex,
@@ -491,7 +487,7 @@ namespace Unity.Rendering
             }
         }
 
-        private void InitializeSphereTest(ref BatchCullingContext cullingContext, ShadowProjection shadowProjection, AllocatorManager.AllocatorHandle allocator)
+        private void InitializeSphereTest(ref BatchCullingContext cullingContext, ShadowProjection shadowProjection)
         {
             // Receiver sphere testing is only enabled if the cascade projection is stable
             bool projectionIsStable = shadowProjection == ShadowProjection.StableFit;
@@ -509,24 +505,17 @@ namespace Unity.Rendering
 
             if (projectionIsStable && allSplitsHaveValidReceiverSpheres)
             {
-                LightAxisX = new float4(cullingContext.localToWorldMatrix.GetColumn(0)).xyz;
-                LightAxisY = new float4(cullingContext.localToWorldMatrix.GetColumn(1)).xyz;
-
-                SplitSOASphereTest = new SOASphereTest(ref this, allocator);
-
+                ReceiverSphereCuller = new ReceiverSphereCuller(cullingContext, this);
                 SphereTestEnabled = true;
             }
         }
-
-        public float2 TransformToLightSpaceXY(float3 positionWS) => new float2(
-            math.dot(positionWS, LightAxisX),
-            math.dot(positionWS, LightAxisY));
     }
 
     internal unsafe struct CullingSplitData
     {
         public float3 CullingSphereCenter;
         public float CullingSphereRadius;
+        public float ShadowCascadeBlendCullingFactor;
         public int PlanePacketOffset;
         public int PlanePacketCount;
         public int CombinedPlanePacketOffset;
@@ -823,6 +812,10 @@ namespace Unity.Rendering
 
             var worldRenderBounds = chunk.GetNativeArray(ref BoundsComponent);
 
+            int visibleSplitMask = ~0;
+            if (useSphereTest)
+                visibleSplitMask = Splits.ReceiverSphereCuller.Cull(chunkBounds.Value);
+
             // First, perform frustum and receiver plane culling for all splits
             for (int splitIndex = 0; splitIndex < splits.Length; ++splitIndex)
             {
@@ -837,10 +830,7 @@ namespace Unity.Rendering
                     s.CombinedPlanePacketOffset,
                     s.CombinedPlanePacketCount);
 
-                float2 receiverSphereLightSpace = Splits.TransformToLightSpaceXY(s.CullingSphereCenter);
-
-                // If the entire chunk fails the sphere test, no need to consider further
-                if (useSphereTest && SphereTest(s, chunkBounds.Value, receiverSphereLightSpace) == SphereTestResult.CannotCastShadow)
+                if ((visibleSplitMask & (1 << splitIndex)) == 0)
                     continue;
 
                 var chunkIn = perInstanceCull
@@ -919,22 +909,7 @@ namespace Unity.Rendering
                         var bounds = worldRenderBounds[entityIndex].Value;
 
                         int planeSplitMask = chunkVisibility->SplitMasks[entityIndex];
-                        int sphereSplitMask = Splits.SplitSOASphereTest.SOASphereTestSplitMask(ref Splits, bounds);
-
-#if DEBUG_VALIDATE_SOA_SPHERE_TEST
-                        int referenceSphereSplitMask = 0;
-                        for (int splitIndex = 0; splitIndex < splits.Length; ++splitIndex)
-                        {
-                            var s = splits[splitIndex];
-                            byte splitMask = (byte)(1 << splitIndex);
-                            float2 receiverSphereLightSpace = Splits.TransformToLightSpaceXY(s.CullingSphereCenter);
-                            if (SphereTest(s, bounds, receiverSphereLightSpace) == SphereTestResult.MightCastShadow)
-                                referenceSphereSplitMask |= splitMask;
-                        }
-                        // Use Debug.Log instead of Debug.Assert so that Burst does not remove it
-                        if (sphereSplitMask != referenceSphereSplitMask)
-                            Debug.Log($"SoA sphere test ({sphereSplitMask:x2}) disagrees with reference sphere tests ({referenceSphereSplitMask:x2})");
-#endif
+                        int sphereSplitMask = Splits.ReceiverSphereCuller.Cull(bounds);
 
                         byte newSplitMask = (byte)(planeSplitMask & sphereSplitMask);
                         chunkVisibility->SplitMasks[entityIndex] = newSplitMask;
@@ -947,106 +922,372 @@ namespace Unity.Rendering
                 }
             }
         }
-
-        private enum SphereTestResult
-        {
-            // The caster is guaranteed to not cast a visible shadow in the tested cascade
-            CannotCastShadow,
-            // The caster might cast a shadow in the tested cascade, and has to be rendered in the shadow map
-            MightCastShadow,
-        }
-
-        private SphereTestResult SphereTest(CullingSplitData split, AABB aabb, float2 receiverSphereLightSpace)
-        {
-            // This test has been ported from the corresponding test done by Unity's
-            // built in shadow culling.
-
-            float casterRadius = math.length(aabb.Extents);
-            float2 casterCenterLightSpaceXY = Splits.TransformToLightSpaceXY(aabb.Center);
-
-            // A spherical caster casts a cylindrical shadow volume. In XY in light space this ends up being a circle/circle intersection test.
-            // Thus we first check if the caster bounding circle is at least partially inside the cascade circle.
-            float sqrDistBetweenCasterAndCascadeCenter = math.lengthsq(casterCenterLightSpaceXY - receiverSphereLightSpace);
-            float combinedRadius = casterRadius + split.CullingSphereRadius;
-            float sqrCombinedRadius = combinedRadius * combinedRadius;
-
-            // If the 2D circles intersect, then the caster is potentially visible in the cascade.
-            // If they don't intersect, then there is no way for the caster to cast a shadow that is
-            // visible inside the circle.
-            // Casters that intersect the circle but are behind the receiver sphere also don't cast shadows.
-            // We don't consider that here, since those casters should be culled out by the receiver
-            // plane culling.
-            if (sqrDistBetweenCasterAndCascadeCenter <= sqrCombinedRadius)
-                return SphereTestResult.MightCastShadow;
-            else
-                return SphereTestResult.CannotCastShadow;
-        }
     }
 
-    internal struct SOASphereTest
+    internal struct ReceiverSphereCuller
     {
-        [NoAlias] public UnsafeList<float4> ReceiverCenterX;
-        [NoAlias] public UnsafeList<float4> ReceiverCenterY;
-        [NoAlias] public UnsafeList<float4> ReceiverRadius;
+        float4 ReceiverSphereCenterX4;
+        float4 ReceiverSphereCenterY4;
+        float4 ReceiverSphereCenterZ4;
+        float4 LSReceiverSphereCenterX4;
+        float4 LSReceiverSphereCenterY4;
+        float4 LSReceiverSphereCenterZ4;
+        float4 ReceiverSphereRadius4;
+        float4 CoreSphereRadius4;
+        UnsafeList<Plane> ShadowFrustumPlanes;
 
-        public SOASphereTest(ref CullingSplits splits, AllocatorManager.AllocatorHandle allocator)
+        float3 LightAxisX;
+        float3 LightAxisY;
+        float3 LightAxisZ;
+        int NumSplits;
+
+        public ReceiverSphereCuller(in BatchCullingContext cullingContext, in CullingSplits splits)
         {
             int numSplits = splits.Splits.Length;
-            int numPackets = (numSplits + 3) / 4;
 
+            Debug.Assert(numSplits <= 4, "More than 4 culling splits is not supported for sphere testing");
             Debug.Assert(numSplits > 0, "No valid culling splits for sphere testing");
 
-            ReceiverCenterX = new UnsafeList<float4>(numPackets, allocator);
-            ReceiverCenterY = new UnsafeList<float4>(numPackets, allocator);
-            ReceiverRadius = new UnsafeList<float4>(numPackets, allocator);
-            ReceiverCenterX.Resize(numPackets);
-            ReceiverCenterY.Resize(numPackets);
-            ReceiverRadius.Resize(numPackets);
+            if (numSplits > 4)
+                numSplits = 4;
 
-            // Initialize the last packet with values that will always fail the sphere test
-            int lastPacket = numPackets - 1;
-            ReceiverCenterX[lastPacket] = new float4(float.PositiveInfinity);
-            ReceiverCenterY[lastPacket] = new float4(float.PositiveInfinity);
-            ReceiverRadius[lastPacket] = float4.zero;
+            // Initialize with values that will always fail the sphere test
+            ReceiverSphereCenterX4 = new float4(float.PositiveInfinity);
+            ReceiverSphereCenterY4 = new float4(float.PositiveInfinity);
+            ReceiverSphereCenterZ4 = new float4(float.PositiveInfinity);
+            LSReceiverSphereCenterX4 = new float4(float.PositiveInfinity);
+            LSReceiverSphereCenterY4 = new float4(float.PositiveInfinity);
+            LSReceiverSphereCenterZ4 = new float4(float.PositiveInfinity);
+            ReceiverSphereRadius4 = float4.zero;
+            CoreSphereRadius4 = float4.zero;
+
+            LightAxisX = new float4(cullingContext.localToWorldMatrix.GetColumn(0)).xyz;
+            LightAxisY = new float4(cullingContext.localToWorldMatrix.GetColumn(1)).xyz;
+            LightAxisZ = new float4(cullingContext.localToWorldMatrix.GetColumn(2)).xyz;
+            NumSplits = numSplits;
+
+            ShadowFrustumPlanes = GetUnsafeListView(cullingContext.cullingPlanes,
+                cullingContext.receiverPlaneOffset,
+                cullingContext.receiverPlaneCount);
 
             for (int i = 0; i < numSplits; ++i)
             {
-                int packetIndex = i >> 2;
                 int elementIndex = i & 3;
+                ref CullingSplitData split = ref splits.Splits.ElementAt(i);
+                float3 lsReceiverSphereCenter = TransformToLightSpace(split.CullingSphereCenter, LightAxisX, LightAxisY, LightAxisZ);
 
-                float2 receiverCenter = splits.TransformToLightSpaceXY(splits.Splits[i].CullingSphereCenter);
-                ReceiverCenterX.ElementAt(packetIndex)[elementIndex] = receiverCenter.x;
-                ReceiverCenterY.ElementAt(packetIndex)[elementIndex] = receiverCenter.y;
-                ReceiverRadius.ElementAt(packetIndex)[elementIndex] = splits.Splits[i].CullingSphereRadius;
+                ReceiverSphereCenterX4[elementIndex] = split.CullingSphereCenter.x;
+                ReceiverSphereCenterY4[elementIndex] = split.CullingSphereCenter.y;
+                ReceiverSphereCenterZ4[elementIndex] = split.CullingSphereCenter.z;
+
+                LSReceiverSphereCenterX4[elementIndex] = lsReceiverSphereCenter.x;
+                LSReceiverSphereCenterY4[elementIndex] = lsReceiverSphereCenter.y;
+                LSReceiverSphereCenterZ4[elementIndex] = lsReceiverSphereCenter.z;
+
+                ReceiverSphereRadius4[elementIndex] = split.CullingSphereRadius;
+                CoreSphereRadius4[elementIndex] = split.CullingSphereRadius * split.ShadowCascadeBlendCullingFactor;
             }
         }
 
-        public int SOASphereTestSplitMask(ref CullingSplits splits, AABB aabb)
+        public int Cull(AABB aabb)
         {
-            int numPackets = ReceiverRadius.Length;
+            int visibleSplitMask = CullSIMD(aabb);
 
-            float4 casterRadius = new float4(math.length(aabb.Extents));
-            float2 casterCenter = splits.TransformToLightSpaceXY(aabb.Center);
-            float4 casterCenterX = casterCenter.xxxx;
-            float4 casterCenterY = casterCenter.yyyy;
+#if DEBUG_VALIDATE_VECTORIZED_CULLING
+            int referenceSplitMask = CullNonSIMD(aabb);
 
-            int splitMask = 0;
-            int splitMaskShift = 0;
-            for (int i = 0; i < numPackets; ++i)
+            // Use Debug.Log instead of Debug.Assert so that Burst does not remove it
+            if (visibleSplitMask != referenceSplitMask)
+                Debug.Log($"Vectorized culling test ({visibleSplitMask:x2}) disagrees with reference test ({referenceSplitMask:x2})");
+#endif
+
+            return visibleSplitMask;
+        }
+
+        int CullSIMD(AABB aabb)
+        {
+            float4 casterRadius4 = new float4(math.length(aabb.Extents));
+            float4 combinedRadius4 = casterRadius4 + ReceiverSphereRadius4;
+            float4 combinedRadiusSq4 = combinedRadius4 * combinedRadius4;
+
+            float3 lsCasterCenter = TransformToLightSpace(aabb.Center, LightAxisX, LightAxisY, LightAxisZ);
+            float4 lsCasterCenterX4 = lsCasterCenter.xxxx;
+            float4 lsCasterCenterY4 = lsCasterCenter.yyyy;
+            float4 lsCasterCenterZ4 = lsCasterCenter.zzzz;
+
+            float4 lsCasterToReceiverSphereX4 = lsCasterCenterX4 - LSReceiverSphereCenterX4;
+            float4 lsCasterToReceiverSphereY4 = lsCasterCenterY4 - LSReceiverSphereCenterY4;
+            float4 lsCasterToReceiverSphereSqX4 = lsCasterToReceiverSphereX4 * lsCasterToReceiverSphereX4;
+            float4 lsCasterToReceiverSphereSqY4 = lsCasterToReceiverSphereY4 * lsCasterToReceiverSphereY4;
+
+            float4 lsCasterToReceiverSphereDistanceSq4 = lsCasterToReceiverSphereSqX4 + lsCasterToReceiverSphereSqY4;
+            bool4 doCirclesOverlap4 = lsCasterToReceiverSphereDistanceSq4 <= combinedRadiusSq4;
+
+            float4 lsZMaxAccountingForCasterRadius4 = LSReceiverSphereCenterZ4 + math.sqrt(combinedRadiusSq4 - lsCasterToReceiverSphereSqX4 - lsCasterToReceiverSphereSqY4);
+            bool4 isBehindCascade4 = lsCasterCenterZ4 <= lsZMaxAccountingForCasterRadius4;
+
+            int isFullyCoveredByCascadeMask = 0b1111;
+
+#if !DISABLE_SHADOW_CULLING_CAPSULE_TEST
+            float3 shadowCapsuleBegin;
+            float3 shadowCapsuleEnd;
+            float shadowCapsuleRadius;
+            ComputeShadowCapsule(LightAxisZ, aabb.Center, casterRadius4.x, ShadowFrustumPlanes,
+                out shadowCapsuleBegin, out shadowCapsuleEnd, out shadowCapsuleRadius);
+
+            bool4 isFullyCoveredByCascade4 = IsCapsuleInsideSphereSIMD(shadowCapsuleBegin, shadowCapsuleEnd, shadowCapsuleRadius,
+                ReceiverSphereCenterX4, ReceiverSphereCenterY4, ReceiverSphereCenterZ4, CoreSphereRadius4);
+
+            if (math.any(isFullyCoveredByCascade4))
             {
-                float4 dx = casterCenterX - ReceiverCenterX[i];
-                float4 dy = casterCenterY - ReceiverCenterY[i];
-                float4 sqrDistBetweenCasterAndCascadeCenter = dx * dx + dy * dy;
-                float4 combinedRadius = casterRadius + ReceiverRadius[i];
-                float4 sqrCombinedRadius = combinedRadius * combinedRadius;
-                bool4 mightCastShadow = sqrDistBetweenCasterAndCascadeCenter <= sqrCombinedRadius;
-                int splitMask4 = math.bitmask(mightCastShadow);
-                // Packet 0 is for bits 0-3, packet 1 is for bits 4-7 etc.
-                splitMask |= splitMask4 << splitMaskShift;
-                splitMaskShift += 4;
+                // The goal here is to find the first non-zero bit in the mask, then set all the bits after it to 0 and all the ones before it to 1.
+
+                // So for example 1100 should become 0111. The transformation logic looks like this:
+                // Find first non-zero bit with tzcnt and build a mask -> 0100
+                // Left shift by one -> 1000
+                // Subtract 1 -> 0111
+
+                int boolMask = math.bitmask(isFullyCoveredByCascade4);
+                isFullyCoveredByCascadeMask = 1 << math.tzcnt(boolMask);
+                isFullyCoveredByCascadeMask = isFullyCoveredByCascadeMask << 1;
+                isFullyCoveredByCascadeMask = isFullyCoveredByCascadeMask - 1;
+            }
+#endif
+
+            return math.bitmask(doCirclesOverlap4 & isBehindCascade4) & isFullyCoveredByCascadeMask;
+        }
+
+        // Keep non-SIMD version around for debugging and validation purposes.
+        int CullNonSIMD(AABB aabb)
+        {
+            // This test has been ported from the corresponding test done by Unity's built in shadow culling.
+
+            float casterRadius = math.length(aabb.Extents);
+
+            float3 lsCasterCenter = TransformToLightSpace(aabb.Center, LightAxisX, LightAxisY, LightAxisZ);
+            float2 lsCasterCenterXY = new float2(lsCasterCenter.x, lsCasterCenter.y);
+
+#if !DISABLE_SHADOW_CULLING_CAPSULE_TEST
+            float3 shadowCapsuleBegin;
+            float3 shadowCapsuleEnd;
+            float shadowCapsuleRadius;
+            ComputeShadowCapsule(LightAxisZ, aabb.Center, casterRadius, ShadowFrustumPlanes,
+                out shadowCapsuleBegin, out shadowCapsuleEnd, out shadowCapsuleRadius);
+#endif
+
+            int visibleSplitMask = 0;
+
+            for (int i = 0; i < NumSplits; i++)
+            {
+                float receiverSphereRadius = ReceiverSphereRadius4[i];
+                float3 lsReceiverSphereCenter = new float3(LSReceiverSphereCenterX4[i], LSReceiverSphereCenterY4[i], LSReceiverSphereCenterZ4[i]);
+                float2 lsReceiverSphereCenterXY = new float2(lsReceiverSphereCenter.x, lsReceiverSphereCenter.y);
+
+                // A spherical caster casts a cylindrical shadow volume. In XY in light space this ends up being a circle/circle intersection test.
+                // Thus we first check if the caster bounding circle is at least partially inside the cascade circle.
+                float lsCasterToReceiverSphereDistanceSq = math.lengthsq(lsCasterCenterXY - lsReceiverSphereCenterXY);
+                float combinedRadius = casterRadius + receiverSphereRadius;
+                float combinedRadiusSq = combinedRadius * combinedRadius;
+
+                // If the 2D circles intersect, then the caster is potentially visible in the cascade.
+                // If they don't intersect, then there is no way for the caster to cast a shadow that is
+                // visible inside the circle.
+                // Casters that intersect the circle but are behind the receiver sphere also don't cast shadows.
+                // We don't consider that here, since those casters should be culled out by the receiver
+                // plane culling.
+                if (lsCasterToReceiverSphereDistanceSq <= combinedRadiusSq)
+                {
+                    float2 lsCasterToReceiverSphereXY = lsCasterCenterXY - lsReceiverSphereCenterXY;
+                    float2 lsCasterToReceiverSphereSqXY = lsCasterToReceiverSphereXY * lsCasterToReceiverSphereXY;
+
+                    // If in light space the shadow caster is behind the current cascade sphere then it can't cast a shadow on it and we can skip it.
+                    // sphere equation is (x - x0)^2 + (y - y0)^2 + (z - z0)^2 = R^2 and we are looking for the farthest away z position
+                    // thus zMaxInLightSpace = z0 + Sqrt(R^2 - (x - x0)^2 - (y - y0)^2 )). R being Cascade + caster radius.
+                    float lsZMaxAccountingForCasterRadius = lsReceiverSphereCenter.z + math.sqrt(combinedRadiusSq - lsCasterToReceiverSphereSqXY.x - lsCasterToReceiverSphereSqXY.y);
+                    if (lsCasterCenter.z > lsZMaxAccountingForCasterRadius)
+                    {
+                        // This is equivalent (but cheaper) than : if (!IntersectCapsuleSphere(shadowVolume, cascades[cascadeIndex].outerSphere))
+                        // As the shadow volume is defined as a capsule, while shadows receivers are defined by a sphere (the cascade split).
+                        // So if they do not intersect there is no need to render that shadow caster for the current cascade.
+                        continue;
+                    }
+
+                    visibleSplitMask |= 1 << i;
+
+#if !DISABLE_SHADOW_CULLING_CAPSULE_TEST
+                    float3 receiverSphereCenter = new float3(ReceiverSphereCenterX4[i], ReceiverSphereCenterY4[i], ReceiverSphereCenterZ4[i]);
+                    float coreSphereRadius = CoreSphereRadius4[i];
+
+                    // Next step is to detect if the shadow volume is fully covered by the cascade. If so we can avoid rendering all other cascades
+                    // as we know that in the case of cascade overlap, the smallest cascade index will always prevail. This help as cascade overlap is usually huge.
+                    if (IsCapsuleInsideSphere(shadowCapsuleBegin, shadowCapsuleEnd, shadowCapsuleRadius, receiverSphereCenter, coreSphereRadius))
+                    {
+                        // Ideally we should test against the union of all cascades up to this one, however in a lot of cases (cascade configuration + light orientation)
+                        // the overlap of current and previous cascades is a super set of the union of these cascades. Thus testing only the previous cascade does
+                        // not create too much overestimation and the math is simpler.
+                        break;
+                    }
+#endif
+                }
             }
 
-            return splitMask;
+            return visibleSplitMask;
+        }
+
+        static void ComputeShadowCapsule(float3 lightDirection, float3 casterPosition, float casterRadius, UnsafeList<Plane> shadowFrustumPlanes,
+            out float3 shadowCapsuleBegin, out float3 shadowCapsuleEnd, out float shadowCapsuleRadius)
+        {
+            float shadowCapsuleLength = GetShadowVolumeLengthFromCasterAndFrustumAndLightDir(lightDirection,
+                casterPosition,
+                casterRadius,
+                shadowFrustumPlanes);
+
+            shadowCapsuleBegin = casterPosition;
+            shadowCapsuleEnd = casterPosition + shadowCapsuleLength * lightDirection;
+            shadowCapsuleRadius = casterRadius;
+        }
+
+        static float GetShadowVolumeLengthFromCasterAndFrustumAndLightDir(float3 lightDir, float3 casterPosition, float casterRadius, UnsafeList<Plane> planes)
+        {
+            // The idea here is to find the capsule that goes from the caster and cover all possible shadow receiver in the frustum.
+            // First we find the distance from the caster center to the frustum
+            var casterRay = new Ray(casterPosition, lightDir);
+            int planeIndex;
+            float distFromCasterToFrustumInLightDirection = RayDistanceToFrustumOriented(casterRay, planes, out planeIndex);
+            if (planeIndex == -1)
+            {
+                // Shadow caster center is outside of frustum and ray do not intersect it.
+                // Shadow volume is thus the caster bounding sphere.
+                return 0;
+            }
+
+            // Then we need to account for the radius of the capsule.
+            // The distance returned might actually be too large in the case of a caster outside of the frustum
+            // however detecting this would require to run another RayDistanceToFrustum and the case is rare enough
+            // so its not a problem (these caster will just be less likely to be culled away).
+            Debug.Assert(planeIndex >= 0 && planeIndex < planes.Length);
+
+            float distFromCasterToPlane = math.abs(planes[planeIndex].GetDistanceToPoint(casterPosition));
+            float sinAlpha = distFromCasterToPlane / (distFromCasterToFrustumInLightDirection + 0.0001f);
+            float tanAlpha = sinAlpha / (math.sqrt(1.0f - (sinAlpha * sinAlpha)));
+            distFromCasterToFrustumInLightDirection += casterRadius / (tanAlpha + 0.0001f);
+
+            return distFromCasterToFrustumInLightDirection;
+        }
+
+        // Returns the shortest distance to the front facing plane from the ray.
+        // Return -1 if no plane intersect this ray.
+        // planeNumber will contain the index of the plane found or -1.
+        static float RayDistanceToFrustumOriented(Ray ray, UnsafeList<Plane> planes, out int planeNumber)
+        {
+            planeNumber = -1;
+            float maxDistance = float.PositiveInfinity;
+            for (int i = 0; i < planes.Length; ++i)
+            {
+                float distance;
+                if (IntersectRayPlaneOriented(ray, planes[i], out distance) && distance < maxDistance)
+                {
+                    maxDistance = distance;
+                    planeNumber = i;
+                }
+            }
+
+            return planeNumber != -1 ? maxDistance : -1.0f;
+        }
+
+        static bool IntersectRayPlaneOriented(Ray ray, Plane plane, out float distance)
+        {
+            distance = 0f;
+
+            float vdot = math.dot(ray.direction, plane.normal);
+            float ndot = -math.dot(ray.origin, plane.normal) - plane.distance;
+
+            // No collision if the ray it the plane from behind
+            if (vdot > 0)
+                return false;
+
+            // is line parallel to the plane? if so, even if the line is
+            // at the plane it is not considered as intersection because
+            // it would be impossible to determine the point of intersection
+            if (Mathf.Approximately(vdot, 0.0F))
+                return false;
+
+            // the resulting intersection is behind the origin of the ray
+            // if the result is negative ( enter < 0 )
+            distance = ndot / vdot;
+
+            return distance > 0.0F;
+        }
+
+        static bool IsInsideSphere(BoundingSphere sphere, BoundingSphere containingSphere)
+        {
+            if (sphere.radius >= containingSphere.radius)
+                return false;
+
+            float squaredDistance = math.lengthsq(containingSphere.position - sphere.position);
+            float radiusDelta = containingSphere.radius - sphere.radius;
+            float squaredRadiusDelta = radiusDelta * radiusDelta;
+
+            return squaredDistance < squaredRadiusDelta;
+        }
+
+        static bool4 IsInsideSphereSIMD(float4 sphereCenterX, float4 sphereCenterY, float4 sphereCenterZ, float4 sphereRadius,
+            float4 containingSphereCenterX, float4 containingSphereCenterY, float4 containingSphereCenterZ, float4 containingSphereRadius)
+        {
+            float4 dx = containingSphereCenterX - sphereCenterX;
+            float4 dy = containingSphereCenterY - sphereCenterY;
+            float4 dz = containingSphereCenterZ - sphereCenterZ;
+
+            float4 squaredDistance = dx * dx + dy * dy + dz * dz;
+            float4 radiusDelta = containingSphereRadius - sphereRadius;
+            float4 squaredRadiusDelta = radiusDelta * radiusDelta;
+
+            bool4 canSphereFit = sphereRadius < containingSphereRadius;
+            bool4 distanceTest = squaredDistance < squaredRadiusDelta;
+
+            return canSphereFit & distanceTest;
+        }
+
+        static bool IsCapsuleInsideSphere(float3 capsuleBegin, float3 capsuleEnd, float capsuleRadius, float3 sphereCenter, float sphereRadius)
+        {
+            var sphere = new BoundingSphere(sphereCenter, sphereRadius);
+            var beginPoint = new BoundingSphere(capsuleBegin, capsuleRadius);
+            var endPoint = new BoundingSphere(capsuleEnd, capsuleRadius);
+
+            return IsInsideSphere(beginPoint, sphere) && IsInsideSphere(endPoint, sphere);
+        }
+
+        static bool4 IsCapsuleInsideSphereSIMD(float3 capsuleBegin, float3 capsuleEnd, float capsuleRadius,
+            float4 sphereCenterX, float4 sphereCenterY, float4 sphereCenterZ, float4 sphereRadius)
+        {
+            float4 beginSphereX = capsuleBegin.xxxx;
+            float4 beginSphereY = capsuleBegin.yyyy;
+            float4 beginSphereZ = capsuleBegin.zzzz;
+
+            float4 endSphereX = capsuleEnd.xxxx;
+            float4 endSphereY = capsuleEnd.yyyy;
+            float4 endSphereZ = capsuleEnd.zzzz;
+
+            float4 capsuleRadius4 = new float4(capsuleRadius);
+
+            bool4 isInsideBeginSphere = IsInsideSphereSIMD(beginSphereX, beginSphereY, beginSphereZ, capsuleRadius4,
+                sphereCenterX, sphereCenterY, sphereCenterZ, sphereRadius);
+
+            bool4 isInsideEndSphere = IsInsideSphereSIMD(endSphereX, endSphereY, endSphereZ, capsuleRadius4,
+                sphereCenterX, sphereCenterY, sphereCenterZ, sphereRadius);
+
+            return isInsideBeginSphere & isInsideEndSphere;
+        }
+
+        static float3 TransformToLightSpace(float3 positionWS, float3 lightAxisX, float3 lightAxisY, float3 lightAxisZ) => new float3(
+            math.dot(positionWS, lightAxisX),
+            math.dot(positionWS, lightAxisY),
+            math.dot(positionWS, lightAxisZ));
+
+        static unsafe UnsafeList<Plane> GetUnsafeListView(NativeArray<Plane> array, int start, int length)
+        {
+            NativeArray<Plane> subArray = array.GetSubArray(start, length);
+            return new UnsafeList<Plane>((Plane*)subArray.GetUnsafeReadOnlyPtr(), length);
         }
     }
 }